telegram-crypto-bot/requirements.txt
APScheduler==3.6.3
beautifulsoup4==4.9.3
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
dnspython==2.1.0
idna==2.10
lxml==4.6.3
numpy==1.21.0
pandas==1.3.0
pymongo==3.11.4
python-dateutil==2.8.1
python-telegram-bot==13.7
pytz==2021.1
requests==2.25.1
six==1.16.0
soupsieve==2.2.1
tornado==6.1
tzlocal==2.1
urllib3==1.26.6
